Add pgm support
This commit adds support for saving and loading image files in the
pgm format. It supports both 8-bit and 16-bit depths; although,
it deviates from the standard by not allowing arbitrary bit depth
values less than the max value of unsigned 16 bit numbers. This
commit includes unit tests.

+++ tests/pbm.rs
@@ -48,7 +48,7 @@ const H:[u8;224] = [1,1,1,1,0,0,0,0,0,0,0,0,1,1,1,1,
                     1,1,1,1,0,0,0,0,0,0,0,0,1,1,1,1,
                     1,1,1,1,0,0,0,0,0,0,0,0,1,1,1,1];
 
-pub fn test_arrs(size:u8, a:&[u8], b:&[u8]) -> bool {
+pub fn test_arrs(size:u32, a:&[u8], b:&[u8]) -> bool {
     let mut ret = true;
     for i in 0..size {
         ret = ret && (a[i as usize] == b[i as usize]);
